fix: drop non-root verifier user; verifier now runs as root
SWEBench tasks require write access during verification — pip install -e .
writes build artifacts into /testbed, setup.py install writes to conda
site-packages — both root-owned after the Tier 4 restore. The non-root
"verifier" OS user blocked universally on these tasks.

Security is carried by content integrity (rsync --delete from /testbed_verify
restores every source file to the trusted pre-agent snapshot before any code
runs) rather than user isolation. All six known attack vectors remain closed
via Tier 2 snapshot/restore, Tier 4 full workspace restore, CLEANUP_CMD, and
VERIFIER_ENV hardening.

- _setup_verifier_user → _seed_verifier_workspace: strips OS user creation
  (groupadd, useradd, home dir wipe, /logs/verifier chown); keeps /logs parent
  lock and /testbed_verify seeding which Tier 4 rsync depends on
- harden_before_verify: drops task.config.verifier.user assignment and
  chmod -R a-w (root bypasses it anyway; pip install needs workspace writes)
- Tasks needing non-root verifier set [verifier] user = "verifier" in task.toml
- test_sandbox_verifier_user.py → test_sandbox_verifier_workspace.py: removes
  5 OS-user creation tests, keeps 3 workspace-seeding tests
- test_sandbox_hardening.py: removes 3 vacuous user-assignment tests, updates
  4 tests to use chown -R root:root as ordering anchor instead of chmod -R a-w
- harden-sandbox.md: Tier 3/4 descriptions, threat table, SDK flow, future dirs
